DimeCoins/requirements/requirements.txt
coinbase==2.0.7
Django==2.0.1
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-debug-toolbar==1.9
django-filter==1.0.4
django-oauth-toolkit==1.0.0
django-two-factor-auth==1.7.0
django-unixdatetimefield==0.1.6
django-user-agents==0.3.2
djangorestframework==3.7.7
djangorestframework-jsonapi==2.3.1
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
enum34==1.1.6
krakenex==2.0.0
django-otp==0.4.2
pytz==2017.3
requests==2.18.4
style==1.1.0
urllib3==1.22
tzlocal==1.5.1